third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/origin-exact-matching/30-expected.txt

CONSOLE WARNING: line 24: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=%22*%22' from origin 'http://localhost:8000' has been blocked by CORS policy: The 'Access-Control-Allow-Origin' header contains the invalid value '"*"'.


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Check that exact matching is used when comparing a request's originating url and the value provided by Access-Control-Allow-Origin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Should disallow origin: '"*"'
PASS xhr.send(null) threw exception NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://127.0.0.1:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=%22*%22'..
PASS successfullyParsed is true

TEST COMPLETE
